Fix issue of select items not being rendered.
When trying out demo, only the first few items were being rendered propery. Checked to see if
it had to do with the loading order of the css packages, but that wasn't it. Demo was fixed
by building them with the demo package. In order to do this I removed the include property to allow
webpack to pull from node_modules.